requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
mysql-connector-python
python-multipart
passlib[bcrypt]
python-jose
python-jose[cryptography]
PyJWT